seperating logic from markup
